Create class ConcurrentLocalSocketListener
